requirements.txt
selenium~=3.141.0
requests~=2.31.0
pytest~=7.4.3
mysql-connector-python~=8.3.0
eyes-selenium~=5.25.2
Faker~=24.4.0
Appium-Python-Client~=1.3.0
urllib3~=1.26.16
webdriver_manager~=3.8.6
allure-pytest~=2.13.2